Moved result to rule conversion logic to new file
